refactor: add TopBlock component, removeTagsFromString function
