+++ src/components/ConstructorGame/ConstructorGame.module.css
@@ -13,108 +13,8 @@
   padding: 5px;
 }
 
-div .word {
-  margin: 0;
-  font-size: 3.6rem;
-  line-height: 4.2rem;
-  position: relative;
-  z-index: 20;
-}
-
-.word__inner {
-  margin: 0 auto;
-}
-
-.description {
-  margin: 5px 0;
-  font-size: 1.4rem;
-  line-height: 1.6rem;
-  text-align: center;
-
-  color: rgba(115, 57, 153, 0.8);
-}
-
-.text {
-  margin: 20px 0;
-  font-size: 1.8rem;
-  line-height: 2.1rem;
-  text-align: center;
-
-  color: rgba(115, 57, 153, 0.8);
-}
-
-.exit-button {
-  display: flex;
-  align-items: center;
-  position: absolute;
-  top: 20px;
-  right: 26px;
-  width: 30px;
-  height: 30px;
-  background-color: inherit;
-  border: none;
-}
-
-.audio-button {
-  display: flex;
-  align-items: center;
-  position: absolute;
-  top: 20px;
-  right: 66px;
-  width: 30px;
-  height: 30px;
-  background-color: inherit;
-  border: none;
-}
-
-.word__container {
-  display: flex;
-  justify-content: center;
-
-  margin-top: 50px;
-  width: 100%;
-}
-
-.word__empty {
-  margin-left: 130px;
-}
-
-.word__transcription {
-  margin: 5px 0;
-  font-size: 1.4rem;
-  line-height: 1.6rem;
-  text-align: center;
-
-  color: rgba(115, 57, 153, 0.8);
-}
-
 .picture {
   width: 250px;
-}
-
-.counter {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
-  margin-right: 60px;
-
-  width: 70px;
-  height: 70px;
-
-  background: linear-gradient(180deg, #ebf3ff 0%, #f7d2fb 100%);
-  border: 5px solid #45f9a3;
-  border-radius: 50px;
-  box-sizing: border-box;
-
-  font-weight: 900;
-  font-size: 1.8rem;
-  line-height: 2.1rem;
-
-  text-align: center;
-  text-transform: capitalize;
-
-  color: #733999;
 }
 
 .cat_sleeping {
@@ -125,26 +25,13 @@ div .word {
   height: 142px;
 }
 
-@media screen and (max-width: 800px) {
-  .word__char {
-    margin: 0 0.4rem;
-  }
-}
-
 @media screen and (max-width: 680px) {
   .cat_sleeping {
     right: -67px;
-  }
-  .word__char {
-    margin: 0 0.2rem;
   }
 }
 @media screen and (max-width: 500px) {
   .cat_sleeping {
     right: -100px;
-  }
-
-  .word__char {
-    margin: 0 0.1rem;
   }
 }
