+++ src/components/ConstructorGame/DragEndDrop/DragEndDrop.module.css
@@ -25,3 +25,20 @@
   text-decoration: none;
   cursor: pointer;
 }
+
+@media screen and (max-width: 800px) {
+  .word__char {
+    margin: 0 0.4rem;
+  }
+}
+
+@media screen and (max-width: 680px) {
+  .word__char {
+    margin: 0 0.2rem;
+  }
+}
+@media screen and (max-width: 500px) {
+  .word__char {
+    margin: 0 0.1rem;
+  }
+}
