+++ src/utils/removeTagsAndWordInside.ts
@@ -1,0 +1,2 @@
+export const removeTagsAndWordInside = (originalString: string) =>
+  originalString.replace(/<([^>]+?)([^>]*?)>(.*?)<\/\1>/gi, '...');
