+++ src/utils/removeTagsFromString.ts
@@ -1,0 +1,2 @@
+export const removeTagsFromString = (originalString: string) =>
+  originalString.replace(/(<([^>]+)>)/gi, '');
